Se colocan enunciados de las consignas del ProyectoFinal
