+++ ProyectoFinal/Controllers/ProductoController.cs
@@ -2,22 +2,26 @@ using ProyectoFinal.Model;
 using ProyectoFinal.Repository;
 using Microsoft.AspNetCore.Mvc;
 using ProyectoFinal.Controllers.DTOS;
+using System.Threading;
 
 namespace ProyectoFinal.Controllers
 {
     [ApiController]
-    //[Route("[Controller]")]
+    [Route("api/[Controller]")]
     public class ProductoController : ControllerBase
     {
+        //Traer Productos: Debe traer todos los productos cargados en la base.
+
         [HttpGet]
-        [Route("api/Producto")]     //Trae TODOS los productos
         public List<Producto> GetAll()
         {
             return ProductoHandler.GetAll();
         }
 
+        //Crear producto: Recibe una lista de tareas por json, número de Id 0, 
+        //Descripción , costo, precio venta y stock.
+
         [HttpPost]
-        [Route("api/Producto")]     //Crea producto
         public bool Create([FromBody] PostProducto producto)
         {
             bool resultado = false;
@@ -39,8 +43,10 @@ namespace ProyectoFinal.Controllers
             return resultado;
         }
         
+        //Modificar producto: Recibe un producto con su número de Id,
+        //debe modificarlo con la nueva información.
+
         [HttpPut]
-        [Route("api/Producto")]     //Modifica producto
         public bool Update([FromBody] PutProducto producto)  
         {
             bool resultado = false;
@@ -65,9 +71,11 @@ namespace ProyectoFinal.Controllers
 
             return resultado;
         }
-        
-        [HttpDelete]
-        [Route("api/Producto/{id}")]    //Elimina producto
+                     
+        //Eliminar producto: Recibe el número de Id de un producto a eliminar
+        //y debe eliminarlo de la base de datos.
+
+        [HttpDelete ("{id}")]
         public bool Delete(int id)
         {
             bool resultado = false;
