+++ ProyectoFinal/Program.cs
@@ -26,6 +26,4 @@ app.UseHttpsRedirection();
 
 app.UseAuthorization();
 
-app.MapControllers();
-
 app.Run();
