+++ ProyectoFinal/Repository/VentaHandler.cs
@@ -35,5 +35,7 @@ namespace ProyectoFinal.Repository
             }
             return resultado; 
         }      
+
+
     }
 }
